progression-command/progression-command-api/src/main/java/uk/gov/moj/cpp/progression/command/InitiateGroupCasesApi.java
package uk.gov.moj.cpp.progression.command;

import static uk.gov.justice.services.core.annotation.Component.COMMAND_API;
import static uk.gov.justice.services.messaging.Envelope.envelopeFrom;
import static uk.gov.justice.services.messaging.Envelope.metadataFrom;

import uk.gov.justice.services.core.annotation.Handles;
import uk.gov.justice.services.core.annotation.ServiceComponent;
import uk.gov.justice.services.core.enveloper.Enveloper;
import uk.gov.justice.services.core.sender.Sender;
import uk.gov.justice.services.messaging.JsonEnvelope;
import uk.gov.justice.services.messaging.Metadata;

import javax.inject.Inject;

@ServiceComponent(COMMAND_API)
public class InitiateGroupCasesApi {

    @Inject
    private Enveloper enveloper;
    @Inject
    private Sender sender;

    @Handles("progression.initiate-court-proceedings-for-group-cases")
    public void handle(final JsonEnvelope envelope) {
        final Metadata metadata = metadataFrom(envelope.metadata())
                .withName("progression.command.initiate-court-proceedings-for-group-cases")
                .build();
        sender.send(envelopeFrom(metadata, envelope.payload()));
    }
}
